Refactor Flask config loading
